Ensure the tests are properly set up
