+++ test/CMakeLists.txt
@@ -1,7 +1,6 @@
 include(owncloud_add_test.cmake)
 
-add_library(syncenginetestutils STATIC testutils/syncenginetestutils.cpp testutils/testutils.cpp)
-target_link_libraries(syncenginetestutils PUBLIC owncloudCore Qt5::Test)
+add_subdirectory(testutils)
 
 owncloud_add_test(OwncloudPropagator)
 owncloud_add_test(Updater)
